create initial data model
