+++ src/extensionConsole/pages/mods/utils/buildGetCanEditModScope.test.ts
@@ -20,7 +20,7 @@ import { modDefinitionFactory } from "@/testUtils/factories/modDefinitionFactori
 import { modMetadataFactory } from "@/testUtils/factories/modComponentFactories";
 import { validateRegistryId } from "@/types/helpers";
 import { organizationStateFactory } from "@/testUtils/factories/authFactories";
-import { UserRole } from "@/types/contract";
+import { UserRole } from "@/data/model/Organization";
 
 const userScope = "my-test-user";
