+++ src/extensionConsole/pages/packageEditor/referenceTab/PackageResult.tsx
@@ -29,7 +29,7 @@ import {
 } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { find as findPackage } from "@/registry/packageRegistry";
-import { type Organization } from "@/types/contract";
+import { type Organization } from "@/data/model/Organization";
 import { type PackageInstance, type Sharing } from "@/types/registryTypes";
 import useAsyncState from "@/hooks/useAsyncState";
