+++ src/types/contract.ts
@@ -40,21 +40,6 @@ import {
 import { type OptionsArgs } from "@/types/runtimeTypes";
 import { type Nullishable } from "@/utils/nullishUtils";
 
-export enum UserRole {
-  member = 1,
-  admin = 2,
-  developer = 3,
-  restricted = 4,
-  manager = 5,
-}
-
-export type Organization = components["schemas"]["Organization"] & {
-  // The `role` property is added in the Redux RTK definition for getOrganizations (see api.ts)
-  // WARNING: currently this role is only accurate for Admin. All other users are passed as Restricted even if they have
-  // a Member or Developer role on the team
-  role: UserRole;
-};
-
 export type Group = components["schemas"]["Group"];
 
 export type Database = components["schemas"]["Database"];
